creation of events and basic editing
